chore: split session sticky handler helper
