+++ sdk/api/handlers/handlers.go
@@ -307,28 +307,6 @@ func requestExecutionMetadata(ctx context.Context) map[string]any {
 	return meta
 }
 
-func requestSessionStickyHeaderKey(c *gin.Context) string {
-	if c == nil {
-		return ""
-	}
-	for _, header := range []string{
-		"Session-Id",
-		"session_id",
-		"X-Session-Id",
-		"X-Codex-Session-Id",
-		"X-Claude-Code-Session-Id",
-		"Conversation-Id",
-		"conversation_id",
-		"X-Conversation-Id",
-		"OpenAI-Conversation-Id",
-	} {
-		if value := strings.TrimSpace(c.GetHeader(header)); value != "" {
-			return "header:" + strings.ToLower(header) + ":" + value
-		}
-	}
-	return ""
-}
-
 func pinnedAuthIDFromContext(ctx context.Context) string {
 	if ctx == nil {
 		return ""
